requirements.txt
asgiref==3.3.1
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
click==8.1.2
click-plugins==1.1.1
cligj==0.7.2
Django==3.1.8
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-gis==0.18
djangorestframework-simplejwt==4.6.0
Fiona==1.8.21
geopandas==0.10.2
#gunicorn==20.1.0
munch==2.5.0
numpy==1.22.3
pandas==1.4.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyproj==3.3.0
python-dateutil==2.8.2
pytz==2021.1
Shapely==1.8.1.post1
six==1.16.0
sqlparse==0.4.1
